ProfileConfigTab: properly handle profile changes
When loading a new profile for a new file source, things like the sample
rate may have changed from the current value that the rest of the UI was
operating on. We need UIMediator to propagate these changes to the
profile throughout the app, so we shouldn't clear the profile changed
flags prematurely before the changes propagated.