Scanner: simplify and fix linear scaling
The old logic didn't handle zoom nicely because scaledPsdAccum was
supposed to represent the full frequency range of the source information
while being at the target bin width. In case of zoom, the source
frequency span is bigger than the target frequency span, so it would cut
off data at its end. This cutting off of data at the end caused the end
of the resulting PSD to always have no data even though the source
contained it.

This new scaling logic is simpler, performs calculations in place
(avoiding the need for the scaled accumulator buffer), and only performs
calculations on the portion of the source frequency range that is within
the target frequency range.

Note that this new implementation no longer interpolates between
frequency bins, for simplicity and performance's sake. Instead, it just
takes the nearest bin. However, I may add back such interpolation if we
later decide it's important enough to justify the performance penalty.