Scanner: allow setting view range before first PSD arrives
In case the user changed the frequency range right before clicking start
on the panoramic spectrum, a signal to update the view range will make
its way here before the first PSD arrives.

I'm keeping the logic to set the spectrum view range based off the
sample rate in fixed frequency mode when the sample rate is known so
that we are not discarding potentially useful data about what's happening
on nearby out-of-view channels when zoomed in. It isn't necessary to do
this (in fact its more efficient to only keep the frequency range we are
currently looking at in the spectrum view), but maybe someone would like
to zoom out and see what was happening on nearby out-of-view frequencies
while they were zoomed in.