Scanner: remove lazy/late initialization
Partially reverts bcb27e300a27c78383a33451aebf6b093c3ec2d2

This was unnecessary, it was previously fixed in suscan with
https://github.com/BatchDrake/suscan/commit/25ce81b844d9017164a667541651278e7f90091d
but that commit had been lost during the switch from the develop to
master branches.

This also fixes a regression that the commit being reverted introduced
breaking zoomed startup of the scanner.